services/scheduler/requirements.txt
APScheduler==3.10.4
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.1
